requirements.txt
blinker==1.9.0
click==8.2.0
dlib==19.24.8
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.1.0
flask-sock==0.7.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0